Refactor code to extract http helper methods into separate module
